preliminary code for plane intersection. Computing above / under plane seems to work

+++ src/plintoperation.h
@@ -38,7 +38,7 @@ private:
     OGRPoint _MultiplyVector (OGRPoint v1, OGRPoint v2);
     bool _ExtractRaster (vrCoordinate * coord);
     void _ComputeABCD (double & a, double & b, double & c, double & d);
-    double _IsUnderOrAbovePlane (const double & a, const double & b, const double & c, const double & d, OGRPoint * pt);
+    float _IsUnderOrAbovePlane (const double & a, const double & b, const double & c, const double & d, OGRPoint * pt);
 };
 
 #endif
